+++ changelog.md
@@ -1,5 +1,6 @@
 
 ## TODO
+	copyMarkup: post-process command list
 	LilyDocument:
 		remove empty track	(6.1 落叶.ly)
 		standalone title: score.header.breakbefore = ##t